% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approximations.R
\name{approx_with_monotonic_spline}
\alias{approx_with_monotonic_spline}
\title{Approximate monotonic spline on data}
\usage{
approx_with_monotonic_spline(bb_response_data, response_var, pred_var,
  env = parent.frame(), increasing, ...)
}
\arguments{
\item{response_var}{Name of response value from bb_response_data.}

\item{pred_var}{Name of predictor value from bb_response_data.}

\item{env}{Formula environment that should be used for fitting gam model.}

\item{increasing}{If TRUE, spline approximation is increasing, if FALSE decreasing.}

\item{...}{Other arguments passed to \link{mgcv::s} function.}

\item{bb_response_data.}{Blackbox response data, for example pdp curve.}
}
\value{
Object of class "gam". See \link{mgcv::gamObject}
}
\description{
It aproximates data with monotonic spline function by fitting GAM model.
}
\examples{
x <- sort(rnorm(20, 5, 5))
y <- rnorm(20, 2, 2)
env <- new.env()
approx_with_monotonic_spline(data.frame(x = x, y = y), "y", "x", env, TRUE)
}
