% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{build_spline_model}
\alias{build_spline_model}
\title{Builds glm model with transformed variables.}
\usage{
build_spline_model(model, data, response_var)
}
\arguments{
\item{model}{Output of \link{DALEX::exlain} function.}

\item{data}{Data used for prediction}
}
\value{
Glm model with transformedd variables with splines.
}
\description{
Builds glm model with transformed variables.
}
