% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-approximations.R
\name{build_xs_function}
\alias{build_xs_function}
\title{It extracts fitted spline function by \link{xspline_approx_gam}.}
\usage{
build_xs_function(quantitative_transition, predictor_name)
}
\arguments{
\item{fitted_gam}{Fitted gam model approximating data.}

\item{...}{Other arguments passed to \link{mgcv::predict.gam} method.}
}
\value{
Function of one variable.
}
\description{
It extracts fitted spline function by \link{xspline_approx_gam}.
}
\examples{
set.seed(123)
x <- rnorm(20, 5, 5)
z <- rnorm(20, 0, 10)
y <- - sin(x) + z ^ 3 + rnorm(20, 0, 0.1)
data <- data.frame(x, y, z)
blackbox <- randomForest::randomForest(y ~ ., data)

z_var_response <- pdp::partial(blackbox, "z")
z_var_response_approx <- approx_with_spline(z_var_response, "yhat", "z")
z_env <- list(
  effect_outcome = z_var_response,
  transition_outcome = z_var_response_approx
)
z_var_spline <- build_xs_function(z_env, "z")
z_range <- attr(z_var_spline, "variable_range")
z_axis <- seq(z_range[1], z_range[2], length.out = 50)

plot(z, y)
lines(z_var_response)
lines(z_axis, z_var_spline(z_axis))

x_var_response <- pdp::partial(blackbox, "x")
x_var_response_approx <- approx_with_spline(x_var_response, "yhat", "x")
x_env <- list(
  effect_outcome = x_var_response,
  transition_outcome = x_var_response_approx
)
x_var_spline <- build_xs_function(x_env, "x")
x_range <- attr(x_var_spline, "variable_range")
x_axis <- seq(x_range[1], x_range[2], length.out = 50)

plot(x, y)
lines(x_var_response)
lines(x_axis, x_var_spline(x_axis))
}
