% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspline.R
\name{build_xspliner}
\alias{build_xspliner}
\title{Helper function for building GLM object with transformed variables.}
\usage{
build_xspliner(formula, model, data, link = "identity",
  family = "gaussian", env = parent.frame(), alter = list(numeric =
  "always", factor = "never"), compare_stat = r_squared)
}
\arguments{
\item{formula}{xspliner-specific formula object. Check vignette("xspliner") for more details.}

\item{model}{Predictive model. Basic model used for extracting predictors transformation.}

\item{data}{Training data of \code{model}.}

\item{link}{Link function that should be used in final model. The passed is used when cannot be extracted from
model. By default 'identity'. See \link{stats::family} for possibilities.}

\item{family}{Family of response variable that should be used in final model. The passed is used when cannot
be extracted from model. By default 'gaussian'. See \link{stats::family} for possibilities.}

\item{env}{Environment in which optional variables passed into parameters are stored.}

\item{alter}{Specifies if each variable transformation should be compared with bare variable usage.
List of the form: list(numeric = type, factor = type), type one of c('always', 'auto', 'never).
'auto' option available only for qualitative variable. The better model is specified by \code{compare_stat} parameter then.}

\item{compare_stat}{Function of linear model (lm function output). Statistic that measures if linear model is better that transformed one.
See \link{stats}.}
}
\description{
Helper function for building GLM object with transformed variables.
}
