% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_helpers.R
\name{get_pdp_spline_improvement}
\alias{get_pdp_spline_improvement}
\title{It builds two glm models with single predictor variable.
The first one is just simple glm model with predictor variable as is.
The second one transforms predictor variable with estimated spline function on the main model response.}
\usage{
get_pdp_spline_improvement(model, data, pred_var, response_var,
  spline_pkg = "gam", type = "pdp", monotonic = NULL, ...)
}
\arguments{
\item{model}{Output of \link{DALEX::explain} function.}

\item{data}{Data used for prediction}

\item{pred_var, }{response_var}

\item{spline_pkg}{Can be "gam" or "rms}

\item{type}{Can be pdp or ale}

\item{...}{Other arguments passed to \link{xspline_approx_gam} method.}
}
\value{
List containing transformation function for respoonse_var, plot on the model and comparison with aic of fitted.
}
\description{
It builds two glm models with single predictor variable.
The first one is just simple glm model with predictor variable as is.
The second one transforms predictor variable with estimated spline function on the main model response.
}
