% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.R
\name{get_xs_call}
\alias{get_xs_call}
\title{It extracts fitted spline function by \link{xspline_approx_gam}.}
\usage{
get_xs_call(xs_env, pred_var_name)
}
\arguments{
\item{fitted_gam}{Fitted gam model approximating data.}

\item{...}{Other arguments passed to \link{mgcv::predict.gam} method.}
}
\value{
Function of one variable.
}
\description{
It extracts fitted spline function by \link{xspline_approx_gam}.
}
\examples{
x <- sort(rnorm(20, 0, 5))
y <- x ^ 3 + rnorm(20, 0, 1)
z <- log(abs(x)) + rnorm(20, 0, 0.01)
data <- data.frame(x, y)
blackbox <- randomForest(y ~ ., data)
x_var_response <- pdp::partial(blackbox, "z")
x_var_response_approx <- approx_with_spline(x_var_response, "yhat", "x")
x_env <- list(
  blackbox_response_obj = x_var_response
  blackbox_response_approx = x_var_response_approx
)
x_var_spline <- get_xs_call(x_env, "x")

gam_model <- xspline_approx_gam(data.frame(x = x, y = y), bs = "tp")
spline <- xspline_function_gam(gam_model)
plot(spline, min(x), max(x))

set.seed(123)
x <- sort(rnorm(40, 0, 5))
y <- sin(x) + log(abs(x)) + rnorm(40, 0, 1)
plot(x, y)
gam_model_mon <- xspline_approx_gam_mon(data.frame(x = x, y = y), increasing = TRUE, bs = "tp")
spline_mon <- xspline_function_gam(gam_model_mon)
gam_model <- xspline_approx_gam(data.frame(x = x, y = y), bs = "tp")
spline <- xspline_function_gam(gam_model)
}
