% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_msg}
\alias{log_msg}
\title{Helper function to print out log messages}
\usage{
log_msg(message)
}
\arguments{
\item{message}{Message that should be printed to R console}
}
\description{
Helper function to print out log messages
}
