% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-xspliner.R
\name{plot.xspliner}
\alias{plot.xspliner}
\title{Plot method for 'xspliner' model}
\usage{
\method{plot}{xspliner}(x, variable_name = NULL, plot_response = TRUE,
  plot_approx = TRUE, data = NULL, plot_data = FALSE, ...)
}
\arguments{
\item{x}{Object of class 'xspliner'}

\item{variable_name}{Name of predictor that should be plotted.}

\item{plot_response}{If TRUE blackbox model response is drawn.}

\item{plot_approx}{If TRUE blackbox model response approcimation is drawn.}

\item{data}{Training data used for building \code{x} model. Required for plot_data option.}

\item{plot_data}{If TRUE raw data is drawn.}
}
\description{
Plot method for 'xspliner' model
}
