% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspline.R
\name{predict.xspliner}
\alias{predict.xspliner}
\title{Predict xspliner method}
\usage{
\method{predict}{xspliner}(object, newdata, ...)
}
\arguments{
\item{object}{Object of class 'xspliner'.}

\item{newdata}{Data that should be prediciton based on.}

\item{...}{Another arguments passed into \link[stats]{predict.glm} method.}
}
\description{
Predict xspliner method
}
