% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-xspliner.R
\name{print.xspliner}
\alias{print.xspliner}
\title{Print method for xspliner object}
\usage{
\method{print}{xspliner}(xspliner, predictor)
}
\arguments{
\item{xspliner}{xspliner object}

\item{predictor}{precitor for xsplner model formula}
}
\description{
Print method for xspliner object
}
