% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspline.R
\name{stats}
\alias{stats}
\alias{r_squared}
\title{Statistics used for better linear model selection}
\usage{
r_squared(lm_model)
}
\arguments{
\item{lm_model}{Linear model - \code{lm} function output.}
}
\description{
Used as \code{compare_stat} parameter in \code{xspline} method.
Each function has attribute "higher-better".
If "higher-better" is TRUE then model with higher statistic value is treated as better one.
}
