% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-xspliner.R
\name{transition}
\alias{transition}
\alias{transition.xspliner}
\title{Extract variable transformation from xspliner}
\usage{
transition(model, ...)

\method{transition}{xspliner}(model, predictor, type = "function", ...)
}
\arguments{
\item{model}{xspliner model}

\item{...}{Other parameters passed to method. Currently not used.}

\item{predictor}{variable name for which transformation should be extracted}

\item{type}{If 'function' then transformation function is extracted. For 'base' there is sourced
object on which transformation was built - in case of quantitative variable GAM model, for qualitative
factorMerger.}
}
\description{
Extract variable transformation from xspliner
}
