% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspline.R
\docType{data}
\name{xf_opts_default}
\alias{xf_opts_default}
\alias{xs_opts_default}
\title{Default parameters for transition methods}
\format{An object of class \code{list} of length 2.}
\usage{
xf_opts_default

xs_opts_default
}
\description{
While constructing formula interpreted by xspliner package, some parameters may be specified within xs(..) or xf(..) symbols.
Below are default parameters. See details in \code{vignette("xspliner")}
}
\keyword{datasets}
