% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_approx}
\alias{xspline_approx}
\title{It aproximates data with spline function by fitting GAM model.}
\usage{
xspline_approx(data, ...)
}
\arguments{
\item{data}{Data with \code{x} and \code{y} variables.}

\item{...}{Other arguments passed to \link{mgcv::s} function.}
}
\value{
Object of class "gam". See \link{mgcv::gamObject}
}
\description{
It aproximates data with spline function by fitting GAM model.
}
\examples{
x <- sort(rnorm(20, 5, 5))
y <- rnorm(20, 2, 2)
xspline_approx(data.frame(x = x, y = y), bs = "tp")
}
