% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_approx_gam_mon}
\alias{xspline_approx_gam_mon}
\title{It aproximates data with spline function by Penalized Constrained Least Squares Fitting.}
\usage{
xspline_approx_gam_mon(data, increasing = TRUE, ...)
}
\arguments{
\item{data}{Data with \code{x} and \code{y} variables.}

\item{increasing}{Should be approximation monotonic}

\item{...}{Other arguments passed to \link{mgcv::s} function.}
}
\value{
Object of class "gam". See \link{mgcv::gamObject}
}
\description{
It aproximates data with spline function by Penalized Constrained Least Squares Fitting.
}
\examples{
x <- sort(rnorm(20, 5, 5))
y <- rnorm(20, 2, 2)
xspline_approx_gam_mon(data.frame(x = x, y = y), bs = "tp", increasing = FALSE)
}
