% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_approx_rms}
\alias{xspline_approx_rms}
\title{It aproximates data with spline function by fitting GAM model with rms package.}
\usage{
xspline_approx_rms(data)
}
\arguments{
\item{data}{Data with \code{x} and \code{y} variables.}

\item{...}{Other arguments passed to \link{mgcv::s} function.}
}
\value{
Object of class "rms".
}
\description{
It aproximates data with spline function by fitting GAM model with rms package.
}
\examples{
x <- sort(rnorm(20, 0, 5))
y <- x^3 + rnorm(20, 0, 1)
rms_model <- xspline_approx_rms(data.frame(x = x, y = y))
}
