% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_function}
\alias{xspline_function}
\title{It extracts fitted spline function by \link{xspline_approx}.}
\usage{
xspline_function(fitted_gam, ...)
}
\arguments{
\item{fitted_gam}{Fitted gam model approximating data.}

\item{...}{Other arguments passed to \link{mgcv::predict.gam} method.}
}
\value{
Function of one variable.
}
\description{
It extracts fitted spline function by \link{xspline_approx}.
}
\examples{
x <- sort(rnorm(20, 5, 5))
y <- rnorm(20, 2, 2)
gam_model <- xspline_approx(data.frame(x = x, y = y), bs = "tp")
spline <- xspline_function(gam_model)
plot(spline)
}
