% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_function_gam}
\alias{xspline_function_gam}
\title{It extracts fitted spline function by \link{xspline_approx_gam}.}
\usage{
xspline_function_gam(fitted_gam, ...)
}
\arguments{
\item{fitted_gam}{Fitted gam model approximating data.}

\item{...}{Other arguments passed to \link{mgcv::predict.gam} method.}
}
\value{
Function of one variable.
}
\description{
It extracts fitted spline function by \link{xspline_approx_gam}.
}
\examples{
x <- sort(rnorm(20, 0, 5))
y <- x^3 + rnorm(20, 0, 1)
gam_model <- xspline_approx_gam(data.frame(x = x, y = y), bs = "tp")
spline <- xspline_function_gam(gam_model)
plot(spline, min(x), max(x))
}
