% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_function_rms}
\alias{xspline_function_rms}
\title{It extracts fitted spline function by \link{xspline_approx}.}
\usage{
xspline_function_rms(fitted_rms)
}
\arguments{
\item{fitted_gam}{Fitted gam model approximating data.}

\item{...}{Other arguments passed to \link{mgcv::predict.gam} method.}
}
\value{
Function of one variable.
}
\description{
It extracts fitted spline function by \link{xspline_approx}.
}
\examples{
x <- sort(rnorm(20, 0, 5))
y <- x^3 + rnorm(20, 0, 1)
rms_model <- xspline_approx_rms(data.frame(x = x, y = y))
spline <- xspline_function_rms(rms_model)
plot(x, y)
plot(spline, min(x), max(x))
}
