% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgcv_approx.R
\name{xspline_plot}
\alias{xspline_plot}
\title{It plots fitted spline function with ggplot package.}
\usage{
xspline_plot(spline_function, add = FALSE, base_plot = NULL)
}
\arguments{
\item{spline_function}{Extracted spline function with 'data' attribute.}

\item{add}{If TRUE the plot is added to another one specified by \code{base_plot} parameter.}

\item{base_plot}{ggplot object to which function should add graph.}
}
\description{
It plots fitted spline function with ggplot package.
}
\examples{
x <- sort(rnorm(20, 5, 5))
y <- rnorm(20, 2, 2)
data <- data.frame(x = x, y = y)
gam_model <- xspline_approx(data, bs = "tp")
spline <- xspline_function(gam_model)
xspline_plot(spline)
p <- ggplot(data, aes(x = x, y = y)) + geom_point()
xspline_plot(spline, add = TRUE, p)
}
