% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-xspliner.R
\name{print.xspliner}
\alias{print.xspliner}
\title{Print method for xspliner object}
\usage{
\method{print}{xspliner}(x, predictor, ...)
}
\arguments{
\item{x}{xspliner object}

\item{predictor}{precitor for xsplner model formula}

\item{...}{Another arguments passed into model specific print method.}
}
\description{
Print method for xspliner object
}
