% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-xspliner.R
\name{summary.xspliner}
\alias{summary.xspliner}
\title{Summary method for xspliner object}
\usage{
\method{summary}{xspliner}(object, predictor, ...)
}
\arguments{
\item{object}{xspliner object}

\item{predictor}{precitor for xspliner model formula}

\item{...}{Another arguments passed into model specific method.}
}
\description{
Summary method for xspliner object
}
