% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-xspliner.R
\name{summary.xspliner}
\alias{summary.xspliner}
\title{Summary method for xspliner object}
\usage{
\method{summary}{xspliner}(xspliner, predictor)
}
\arguments{
\item{xspliner}{xspliner object}

\item{predictor}{precitor for xsplner model formula}
}
\description{
Summary method for xspliner object
}
