% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspline.R
\name{xspline}
\alias{xspline}
\alias{xspline.default}
\alias{xspline.formula}
\alias{xspline.explainer}
\title{Builds predictive model based GLM.}
\usage{
xspline(object, ...)

\method{xspline}{default}(object, lhs = NULL, response = NULL,
  predictors = NULL, data = NULL, form = "additive",
  env = parent.frame(), ...)

\method{xspline}{formula}(object, model, data = NULL,
  consider = "specials", env = parent.frame(), ...)

\method{xspline}{explainer}(object, env = parent.frame(), ...)
}
\arguments{
\item{object}{Predictive model, formula or explainer (see DALEX) object.}

\item{...}{Other arguments passed to \code{xspline} methods or \link{build_xspliner}.}

\item{lhs}{Left-hand side of model formula. Can be transformed response.}

\item{response}{Name of response variable of \code{model}.}

\item{predictors}{Predictor values that should be used in final model.}

\item{data}{Training data of \code{model}.}

\item{form}{Can be 'additive' (default) or 'multiplicative'. Specifies formula form in final model.}

\item{env}{Environment in which optional variables passed into parameters are stored.
variables transformation. See vignette("xspliner") for details.}

\item{model}{When \code{object} is formula - predictive model. Basic model used for extracting predictors transformation.}

\item{consider}{One of \code{c("specials", "all")}. If "specials", only components with xs or xf
call are considered in transition.}
}
\value{
GLM object of class 'xspliner'.
}
\description{
See details in \code{vignette("xspliner")}.
}
