% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspline.R
\name{xspline}
\alias{xspline}
\alias{xspline.default}
\alias{xspline.formula}
\alias{xspline.explainer}
\title{Builds predictive model based GLM.}
\usage{
xspline(object, ...)

\method{xspline}{default}(model, response = NULL, predictors = NULL,
  data = NULL, env = parent.frame(),
  factor_opts = factor_opts_default,
  numeric_opts = numeric_opts_default, ...)

\method{xspline}{formula}(formula, model, data = NULL, exact = TRUE,
  env = parent.frame(), ...)

\method{xspline}{explainer}(explainer, env = parent.frame(), ...)
}
\arguments{
\item{object}{Predictive model, formula or explainer (see \link{DALEX}) object.}

\item{...}{Other arguments passed to \code{xspline} methods or \link{build_xspliner}.}

\item{model}{Predictive model. Basic model used for extracting predictors transformation.}

\item{response}{Name of response variable of \code{model}.}

\item{predictors}{Predictor values that should be used in final model.}

\item{data}{Training data of \code{model}.}

\item{env}{Environment in which optional variables passed into parameters are stored.}

\item{factor_opts}{Formula parameters used for factor variable transformatoins inherited from factorMerger package.}

\item{numeric_opts}{Predictive model response method and approximation parameters used for numeric
variables transformation. See vignette("xspliner") for details.}

\item{formula}{xspliner-specific formula object. Check vignette("xspliner") for more details.}

\item{exact}{If TRUE, exact formula call is used. If not all formula variables are altered.}

\item{explainer}{Object of class 'explainer' (see \link{DALEX} package).}
}
\value{
GLM object of class 'xspliner'.
}
\description{
See details in \code{vignette("xspliner")}.
}
